Updated to use the dataset API where endpoints have been moved from the import API.
